+++ apps/vben5/apps/app-antd/src/views/permissions/groups/index.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import { Page } from '@vben/common-ui';
 
-import { PermissionGroupDefinitionTable } from '@abp/permission';
+import { PermissionGroupDefinitionTable } from '@abp/permissions';
 
 defineOptions({
   name: 'PermissionGroupDefinitions',
