+++ apps/vben5/packages/@abp/identity/package.json
@@ -22,7 +22,7 @@
   "dependencies": {
     "@abp/auditing": "workspace:*",
     "@abp/core": "workspace:*",
-    "@abp/permission": "workspace:*",
+    "@abp/permissions": "workspace:*",
     "@abp/request": "workspace:*",
     "@abp/ui": "workspace:*",
     "@ant-design/icons-vue": "catalog:",
