+++ apps/vben5/packages/@abp/identity/src/components/organization-units/OrganizationUnitTree.vue
@@ -12,7 +12,7 @@ import { useVbenModal } from '@vben/common-ui';
 import { createIconifyIcon } from '@vben/icons';
 import { $t } from '@vben/locales';
 
-import { PermissionModal } from '@abp/permission';
+import { PermissionModal } from '@abp/permissions';
 import {
   DeleteOutlined,
   EditOutlined,
