+++ apps/vben5/packages/@abp/identity/src/components/roles/RoleTable.vue
@@ -13,7 +13,7 @@ import { $t } from '@vben/locales';
 
 import { AuditLogPermissions, EntityChangeDrawer } from '@abp/auditing';
 import { useAbpStore } from '@abp/core';
-import { PermissionModal } from '@abp/permission';
+import { PermissionModal } from '@abp/permissions';
 import { useVbenVxeGrid } from '@abp/ui';
 import {
   DeleteOutlined,
