+++ apps/vben5/packages/@abp/openiddict/package.json
@@ -22,7 +22,7 @@
   "dependencies": {
     "@abp/core": "workspace:*",
     "@abp/identity": "workspace:*",
-    "@abp/permission": "workspace:*",
+    "@abp/permissions": "workspace:*",
     "@abp/request": "workspace:*",
     "@abp/ui": "workspace:*",
     "@ant-design/icons-vue": "catalog:",
